+++ NEWS.md
@@ -5,8 +5,8 @@ This version of `tern` introduces a major rewriting of `tern` due to the change 
 * Counting patients with abnormal values post-baseline: `s_count_abnormal` and `analyze_abnormal`.
 * Summarize variables: S3 generic `s_summary` (with numeric, factor and character methods), `summarize_vars`.
 * Estimate difference in terms of responder proportions: `estimate_proportion_diff`, `s_proportion_diff`.
-* Test the difference in responder proportions: `prop_chisq`, `prop_cmh`.
 * Estimate difference in terms of Odds Ratio: `s_odds_ratio`, `or_glm`.
+* Test the difference in responder proportions: `s_test_proportion_diff`, `prop_chisq`, `prop_cmh`, `prop_fisher`, `prop_schouten`.
 
 * issues with no news: